+++ ts/BoardRenderer.ts
