+++ ts/TrainRenderer.ts
